Framework

Documentation

toAddrString :: (Word8, Word8, Word8, Word8) -> String

gamePort :: PortNumber

chatPort :: PortNumber

Current list of players, may or may not be upto dated depending on the distance from you Current list of food blocks, it will not be upto dated for the block where the block you are not in. Action.

type GameAngle

Arguments

= Word16
Must be in range 0 ~ 4096

type DashFlag

Arguments

= Bool
True if dash is on, False otherwise

type GameReceiveCallback = Index -> [(Index, Player)] -> [(Index, FoodBlock)] -> (DashFlag, GameAngle)

type MutablePlayerList = [(Index, IORef Player)]

type MutableFoodList = [(Index, IORef FoodBlock)]

type PlayerID = IORef Index

pullIORefs :: [(Index, IORef a)] -> IO [(Index, a)]

eliminateIORef :: (Index, IORef a) -> IO (Index, a)

parsePlayer :: MutablePlayerList -> Index -> [Word8] -> IO ()

parseAction :: MutablePlayerList -> Index -> [Word8] -> IO ()

parseDeath :: MutablePlayerList -> Index -> [Word8] -> IO ()

parseFood :: MutableFoodList -> Index -> [Word8] -> IO ()

parseNumber :: IORef Index -> Word8 -> IO ()

A tag for parsing data

playerParseTag :: Word8

actionParseTag :: Word8

foodParseTag :: Word8

deathParseTag :: Word8

numberParseTag :: Word8

parseGameData :: PlayerID -> MutablePlayerList -> MutableFoodList -> [Word8] -> IO ()