Mimizu

Description

Documentation

module Mimizu.Player

module Mimizu.Util

module Mimizu.Food

module Mimizu.Chat

nearestFood :: Player -> [FoodBlock] -> Food

Returns the food in the block which the player is in, in mesh coordinate.

nearestWorldFood :: Player -> [FoodBlock] -> WorldFood

Returns the minimum distance to a food from the player, in world coordinate.

worldFoodDistance :: Player -> WorldFood -> Float

Calculate relational distance to the player in world coordinate.

foodDistance :: Player -> Food -> Float

Warning: This function does not return the correct value when the block tha food is contained by is different from the player is in.

Calculate relational distance to the player in mesh coordinate.

compareByWorldFoodDistance :: Player -> WorldFood -> WorldFood -> Ordering

Compareing the relational distance to the player specified using worldFoodDistance nearer comes first, further comes second.

compareByFoodDistance :: Player -> Food -> Food -> Ordering

Compareing the relational distance to the player specified using foodDistance nearer comes first, further comes second.

sortByWorldFoodDistance :: Player -> [WorldFood] -> [WorldFood]

Sort foods by relational distance to the player, using compareByWorldFoodDistance.

sortByFoodDistance :: Player -> [Food] -> [Food]

Sort foods by relational distance to the player, using foodDistance.

sortByWorldFoodDistanceFromFoodBlock :: Player -> Index -> FoodBlock -> [WorldFood]

Similar to sortByFoodDistanceFromFoodBlock, but this function also takes a block index, returns in world coordinate.

sortByFoodDistanceFromFoodBlock :: Player -> FoodBlock -> [Food]

The same as sortByFoodDistance, but this function takes FoodBlock insted of a list of Food