Mimizu.Chat

Description

Documentation

data Chat

The representation of one message

Constructors

Chat

Fields

origin :: Origin

The origin of message

time :: UTCTime

The time (Internally in UTC, Externally in JST)

sender :: String

The display name of sender the message sent by

message :: String

The actual message

data Origin

The representation of origin where a message come from

Constructors

NicoNico
NicoNico live stream. See http://com.nicovideo.jp/community/co3265652

TUGame
TUGame original site. See http://tk2-217-18218.vs.sakura.ne.jp

YouTube
YouTube live stream. See https://www.youtube.com/user/TsUmezawa

toOriginString :: Origin -> String

convert origin to chat websocket acceptable string

fromOriginString :: String -> Origin

convert origin string to its Enum value