Mimizu.Food

Description

Documentation

newtype FoodBlock

The food info contained by one block.

Constructors

FoodBlock

Fields

foodInfo :: [Word8]

getColor :: Index -> FoodBlock -> Color

color to be displayed

getSize :: Index -> FoodBlock -> Word8

size you can aquire when you eat, also size to be displayed

getFoodX :: Index -> FoodBlock -> Word8

x location of the food specified by index and food block

getFoodY :: Index -> FoodBlock -> Word8

y location of the food specified by index and food block

getX :: Index -> FoodBlock -> Word8

compatibility funcction for getFoodX and getFoodY

getY :: Index -> FoodBlock -> Word8

compatibility funcction for getFoodX and getFoodY