Mimizu.Player

Description

Documentation

type Length = Word32

The type of the length

data Player

The representation of player

Constructors

Player

Fields

skin :: [Color]

Skin information, a sequence of color

name :: String

Name to be displayed

excreta :: Int

Counter to be shorten

act :: Word16

Recent action

playerInfo :: [Word8]

Joints and other info

createPlayer :: [Word8] -> Player

A constructor-like function, compatibility for JS original source

getVA :: Player -> Word32

The length and an air value of the player specified

getAct :: Player -> Word16

The most recent action taken by the player specified

getSBM :: Player -> Word16

The current speed and `maximum speed` for the player specified

getUSX :: Player -> Word16

The rough x-location, this function is only available when the distance between you and the player specified is enough

getUSY :: Player -> Word16

The rough y-location, this function is only available when the distance between you and the player specified is enough

getAir :: Player -> Word32

the counter when the player specified is overlapping/overlapped by other player. when this counter is enough, `going straight` bug will occur.

getAn :: Player -> Word16

The angle being aimed by the player specified

getBtn :: Player -> Bool

The flag whether dash button is been pressed by the player specified, true if pressed

getJN :: Player -> Int32

Number of joints of the player specified

getSH :: Player -> Float

The gear the player currently in

getSpd :: Player -> Word16

Current speed of the player specified

getVol :: Player -> Word32

The actual length the specified player has

getSR :: Player -> Float

The thickness of the player.

getTR :: Player -> Float

Minimum turn radius

getX0 :: Player -> Either Word16 Float

The x location of head of the player specified. if it is too far, return rough value by Left otherwise, return precise value by Right

getY0 :: Player -> Either Word16 Float

The y location of head of the player specified. if it is too far, return rough value by Left otherwise, return precise value by Right

getAngle :: Player -> Float

Current Actual Head the player specified is going

getJointX :: Index -> Player -> Float

The x location of idx-th joints of the player specified

getJointY :: Index -> Player -> Float

The y location of index-th joints of the player specified

sizeR :: Length -> Float

Size convert function from the length

getX :: Index -> Player -> Float

Compatibility function for getJointX and getJointY

getY :: Index -> Player -> Float

Compatibility function for getJointX and getJointY