Framework

Description

Documentation

hostAddress :: (Word8, Word8, Word8, Word8)

The destination IP address the socket connects to, in the form of Word8 quadruplet.

toAddrString :: (Word8, Word8, Word8, Word8) -> String

Convert Word8 quadruple to host string.

gamePort :: PortNumber

The port number the socket accepts.

chatPort :: PortNumber

The port number the socket accepts.

type GameReceiveCallback

Arguments

= Index
The index of player you are currently playing.

-> [(Index, Maybe Player)]
Current list of players, which may not be updated depending on the distance from you.

-> [(Index, FoodBlock)]
Current list of food blocks, which may not be updated for the block where the block you are not in.

-> (DashFlag, GameAngle)
The action you intend to be taken.

The callback function, this is called when message from gamePort.

type GameAngle

Arguments

= Word16
Must be in range 0 ~ 4096.

type DashFlag

Arguments

= Bool
True if dash is on, False otherwise.

type MutablePlayerList = [(Index, IORef (Maybe Player))]

A list/map of players, Nothing if the player for that Index is not being played/dead.

type MutableFoodList = [(Index, IORef FoodBlock)]

A list/map of foods, Index represents the block.

type PlayerID = IORef Index

Your ID for playing, the index is the key of MutablePlayerList.

convertToSendable :: (DashFlag, GameAngle) -> ByteString

Convert to websocket acceptable form.

run :: GameReceiveCallback -> ClientApp ()

mainLoop :: GameReceiveCallback -> IO ()

pullIORefs :: [(Index, IORef a)] -> IO [(Index, a)]

Convert mutable list/map to immutable

eliminateIORef :: (Index, IORef a) -> IO (Index, a)

Remove IORef out of the argument.

parsePlayer :: MutablePlayerList -> Index -> [Word8] -> IO ()

Modify player list using parsed player other info for the specific plaeyr

parseAction :: MutablePlayerList -> Index -> [Word8] -> IO ()

Modify player list using pased action binary data for the specific player

parseDeath :: MutablePlayerList -> Index -> IO ()

Modify player list using parsed death info

parseFood :: MutableFoodList -> Index -> [Word8] -> IO ()

Modify food list using parsed food info.

parseNumber :: IORef Index -> Word8 -> IO ()

Modify ist using parsed index value

playerParseTag :: Word8

A tag for parsing data.

actionParseTag :: Word8

A tag for parsing data.

foodParseTag :: Word8

A tag for parsing data.

deathParseTag :: Word8

A tag for parsing data.

numberParseTag :: Word8

A tag for parsing data.

parseGameData :: PlayerID -> MutablePlayerList -> MutableFoodList -> [Word8] -> IO ()

parse binary data from gameSocket.

parseSkinData :: MutablePlayerList -> Text -> IO ()

parse text data from gameSocket