Framework

Description

Documentation

hostAddress :: (Word8, Word8, Word8, Word8)

The destination IP address the socket connects to, in the form of Word8 quadruplet.

gamePort :: PortNumber

The port number the socket accepts.

chatPort :: PortNumber

The port number the socket accepts.

type GameReceiveCallback

Arguments

= Index
The index of player you are currently playing.

-> [(Index, Maybe Player)]
Current list of players, which may not be updated depending on the distance from you.

-> [(Index, FoodBlock)]
Current list of food blocks, which may not be updated for the block where the block you are not in.

-> IO (DashFlag, GameAngle)
The action you intend to be taken.

The callback function, this is called when message received from gamePort.

type ErrorHandler

Arguments

= Index
The index of player you are currently playing.

-> [(Index, Maybe Player)]
Current list of players, which may not be updated depending on the distance from you.

-> [(Index, FoodBlock)]
Current list of food blocks, which may not be updated for the block where the block you are not in.

-> String
The message of error thrown

-> IO ()

Handler.

type ChatCallback

Arguments

= [Chat]
Chats recieved, including the latest

-> Ranking
The ranking

-> IO Text
The message you want to send, if it is empty sending will not be fired.

The callback function, this is called when message received from chatPort

type GameAngle = Word16

Must be in range 0 ~ 4095.

type DashFlag = Bool

True if dash is on, False otherwise.

type MutablePlayerList = [(Index, IORef (Maybe Player))]

A list/map of players, Nothing if the player for that Index is not being played/dead.

type MutableFoodList = [(Index, IORef FoodBlock)]

A list/map of food blocks.

type PlayerID = IORef Index

Your ID for playing, the index is the key of MutablePlayerList.

type Ranking = [(String, Word32)]

High score of the day (when you started to play)

convertToSendable :: (DashFlag, GameAngle) -> ByteString

Convert to websocket acceptable form.

pullIORefs :: [(Index, IORef a)] -> IO [(Index, a)]

Convert mutable list/map to immutable.

eliminateIORef :: (Index, IORef a) -> IO (Index, a)

Remove IORef out of the argument.

fromRadian :: Double -> GameAngle

Convert radians (which must be in range (-pi, pi].

mainLoop

Arguments

:: String
The session id you want to play.

-> ErrorHandler
The callback, which will be called when error occured.

-> GameReceiveCallback
The callback, which will be called when any new message from game received.

-> ChatCallback
The callback, which will be called when any new message from chat received.

-> ChatCallback
The function, which will be called every 1s, so that you can actively send.

-> Bool
The flag, True if run this thread again, False otherwise.

-> IO ()

Entry point.