Mimizu.Food

Description

Documentation

newtype FoodBlock

The food info contained by one block.

Constructors

FoodBlock

Fields

foodInfo :: [Word8]

type Food = (Word8, Word8)

Tuple of x and y coordinates for specific food in the food block.

type WorldFood = (Word16, Word16)

Tuple of x and y coordinates for specific food in the world coordinate.

color :: Index -> FoodBlock -> Color

Color to be displayed.

size :: Index -> FoodBlock -> Word8

Size you can aquire when you eat, also size to be displayed.

foodX :: Index -> FoodBlock -> Word8

X location of the food specified by index and food block.

foodY :: Index -> FoodBlock -> Word8

Y location of the food specified by index and food block.

x :: Index -> FoodBlock -> Word8

Compatibility funcction for foodX and foodY

y :: Index -> FoodBlock -> Word8

Compatibility funcction for foodX and foodY

foodCount :: FoodBlock -> Index

The number of foods contained by a food block.

foodWorldX :: Index -> Index -> FoodBlock -> Word16

Convert a food to world coordinate specified in first index-th block, second index-th food in food block.

foodWorldY :: Index -> Index -> FoodBlock -> Word16

Convert a food to world coordinate specified in first index-th block, second index-th food in food block.

worldX :: Index -> Index -> FoodBlock -> Word16

Compatibility funcction for foodWorldX and foodWorldY.

worldY :: Index -> Index -> FoodBlock -> Word16

Compatibility funcction for foodWorldX and foodWorldY.

meshPositions :: FoodBlock -> [Food]

Convert food block to list of Foods.

worldPositions :: Index -> FoodBlock -> [WorldFood]

Convert food block to list of WorldFoods.