Mimizu.Player

Description

Documentation

type Length = Word32

The type of the length

data Player

The representation of player

Constructors

Player

Fields

skin :: [Color]

Skin information, a sequence of color

name :: String

Name to be displayed

excreta :: Word32

Counter to be shorten

act :: Word16

Recent action

playerInfo :: [Word8]

Joints and other info

createPlayer :: [Word8] -> Player

A constructor-like function, compatibility for JS original source

va :: Player -> Word32

The length and an air value of the player specified

sbm :: Player -> Word16

The current speed and `maximum speed` for the player specified

usx :: Player -> Word16

The rough x-location, this function is only available when the distance between you and the player specified is enough

usy :: Player -> Word16

The rough y-location, this function is only available when the distance between you and the player specified is enough

air :: Player -> Word32

the counter when the player specified is overlapping/overlapped by other player. when this counter is enough, `going straight` bug will occur.

an :: Player -> Word16

The angle being aimed by the player specified

btn :: Player -> Bool

The flag whether dash button is been pressed by the player specified, true if pressed

jn :: Player -> Int32

Number of joints of the player specified

sh :: Player -> Float

The gear the player currently in

spd :: Player -> Word16

Current speed of the player specified

vol :: Player -> Word32

The actual length the specified player has

sr :: Player -> Float

The thickness of the player.

tr :: Player -> Float

Minimum turn radius

x0 :: Player -> Either Word16 Float

The x location of head of the player specified. if it is too far, return rough value by Left otherwise, return precise value by Right

y0 :: Player -> Either Word16 Float

The y location of head of the player specified. if it is too far, return rough value by Left otherwise, return precise value by Right

angle :: Player -> Float

Current Actual Head the player specified is going

jointX :: Index -> Player -> Float

The x location of idx-th joints of the player specified

jointY :: Index -> Player -> Float

The y location of index-th joints of the player specified

sizeR :: Length -> Float

Size convert function from the length

x :: Index -> Player -> Float

Compatibility function for joinxX and jointY

y :: Index -> Player -> Float

Compatibility function for joinxX and jointY

modifyPlayerInfo :: [Word8] -> Maybe Player -> Maybe Player

The helper function for MutablePlayerList, it will modify the player info.

modifyAction :: Word16 -> Maybe Player -> Maybe Player

The helper function for MutablePlayerList, it will modify the action.

modifyExcreta :: Word32 -> Maybe Player -> Maybe Player

The helper function for MutablePlayerList, it will modify the excreta

modifyName :: String -> Maybe Player -> Maybe Player

The helper function for MutablePlayerList, it will modify the name

modifySkin :: [Color] -> Maybe Player -> Maybe Player

The helper function for MutablePlayerList, it will modify the skin

playerWorldX :: Player -> Word16

Returns X location in world coordinate of specified Player

playerWorldY :: Player -> Word16

Returns X location in world coordinate of specified Player

playerWorldXf :: Player -> Float

Returns X location in world coordinate of specified Player. This function is more precise than worldX, worldY

playerWorldYf :: Player -> Float

Returns X location in world coordinate of specified Player. This function is more precise than worldX, worldY

worldX :: Player -> Word16

Compatibility function for playerWorldX and plaeyrWorldY

worldY :: Player -> Word16

Compatibility function for playerWorldX and plaeyrWorldY

worldXf :: Player -> Float

Compatibility function for playerWorldXf and plaeyrWorldYf

worldYf :: Player -> Float

Compatibility function for playerWorldXf and plaeyrWorldYf

headDistance :: Player -> Player -> Word64

Returns the distance between players, the distance is measured in World coordinate.

headDistancef :: Player -> Player -> Float

Returns the distance between players, the distance is measured in World coordinate. This function is more precise than headDistance

compareByRanking :: Player -> Player -> Ordering

Returns LT if first player's volume is less than the others' GT when greater than, EQ otherwise.

sortByRanking :: [Player] -> [Player]

Sorting playrs given, fits to the ranking.